ref: move o diretório pages para o lugar correto
Seguindo o padrão mais usado na comunidade do React
